+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- system: clean up of outdated dagster results
 - implement AssetCheck `x_percent_less_than`
 
 ### Changes